+++ src/control_plane/migration/mysql/mod.rs
@@ -1,5 +1,9 @@
 mod backup_mysql_database;
 mod mysql_backup_options;
+mod mysql_restore_options;
+mod restore_mysql_database;
 
 pub(crate) use backup_mysql_database::backup_mysql_database;
 pub(crate) use mysql_backup_options::MySqlBackupOptions;
+pub(crate) use mysql_restore_options::MySqlRestoreOptions;
+pub(crate) use restore_mysql_database::restore_mysql_database;